refactor(aletheia): extract runtime::validate submodule (#2642)
Move validate_jwt and validate_external_tools to a dedicated file.
These are used by RuntimeBuilder::validate for config checking.
